+++ src/main/java/com/InvoicifyPro/InvoicifyPro/service/expense/ExpenseServiceImpl.java
@@ -46,6 +46,9 @@ public class ExpenseServiceImpl implements ExpenseService {
     @Override
     @Transactional
     public void deleteById(Long id) {
+        if(expenseRepository.existsById(id)){
+            throw new ResourceNotFoundException("Expense",id );
+        }
         expenseRepository.deleteById(id);
     }
 }
